requirements.txt
about-time==4.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alive-progress==3.2.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
Authlib==1.6.0
backports-datetime-fromisoformat==2.0.3
backports.tarfile==1.2.0
black==25.1.0
boolean.py==5.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
crashtest==0.4.1
cryptography==45.0.5
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
dependency-injector==4.48.1
distlib==0.3.9
dparse==0.6.4
dulwich==0.22.8
exceptiongroup==1.3.0
fastjsonschema==2.21.1
filelock==3.16.1
findpython==0.6.3
flake8==7.3.0
frozenlist==1.7.0
grapheme==0.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jeepney==0.9.0
Jinja2==3.1.6
joblib==1.5.1
keyring==25.6.0
license-expression==30.4.3
linkify-it-py==2.0.3
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
more-itertools==10.7.0
msgpack==1.1.1
multidict==6.6.3
mypy==1.16.1
mypy_extensions==1.1.0
nltk==3.9.1
nodeenv==1.9.1
packageurl-python==0.17.1
packaging==25.0
pathspec==0.12.1
pbs-installer==2025.7.8
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
poetry==2.1.3
poetry-core==2.1.3
propcache==0.3.2
psutil==6.1.1
py-serializable==2.0.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
pyright==1.1.403
pytest==8.4.1
python-dotenv==1.1.1
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safety==3.6.0
safety-schemas==0.0.14
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==9.1.2
textual==3.7.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.67.1
trove-classifiers==2025.5.9.12
typer==0.16.0
types-requests==2.32.4.20250611
typing_extensions==4.14.1
uc-micro-py==1.0.3
urllib3==2.5.0
virtualenv==20.31.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
